Inline `MergedDescriptorSet`
